setup.py
import setuptools
from decouple import config

with open(file="README.md", mode="r", encoding="utf-8") as f:
    long_description = f.read()


__version__ = "0.0.0"

REPO_NAME = "chest-cancer-classification-mlflow-dvc"
AUTHOR_USERNAME = config("GITHUB_USERNAME")
SRC_REPO = "cnnClassifier"
AUTHOR_EMAIL = config("GITHUB_EMAIL")


setuptools.setup(
    name=SRC_REPO,
    version=__version__,
    author=AUTHOR_USERNAME,
    author_email=AUTHOR_EMAIL,
    description="A small python package for CNN app",
    long_description=long_description,
    long_description_content="text/markdown",
    url=f"https://github.com/{AUTHOR_USERNAME}/{REPO_NAME}",
    project_urls={
        "Bug Tracker": f"https://github.com/{AUTHOR_USERNAME}/{REPO_NAME}/issues",
    },
    package_dir={"": "src"},
    packages=setuptools.find_packages(where="src")
)
